adjust path for GIL Refs deprecation warnings
