+++ src/impl_/pymethods.rs
@@ -580,52 +580,3 @@ pub unsafe fn tp_new_impl<T: PyClass>(
         .create_class_object_of_type(py, target_type)
         .map(Bound::into_ptr)
 }
-
-pub fn inspect_type<T>(t: T) -> (T, Extractor<T>) {
-    (t, Extractor::new())
-}
-
-pub struct Extractor<T>(NotAGilRef<T>);
-pub struct NotAGilRef<T>(std::marker::PhantomData<T>);
-
-pub trait IsGilRef {}
-
-impl<T: crate::PyNativeType> IsGilRef for &'_ T {}
-
-impl<T> Extractor<T> {
-    #[allow(clippy::new_without_default)]
-    pub fn new() -> Self {
-        Extractor(NotAGilRef(std::marker::PhantomData))
-    }
-}
-
-impl Extractor<Python<'_>> {
-    #[cfg_attr(
-        not(feature = "gil-refs"),
-        deprecated(since = "0.21.0", note = "use `wrap_pyfunction_bound!` instead")
-    )]
-    pub fn is_python(&self) {}
-}
-
-impl<T: IsGilRef> Extractor<T> {
-    #[cfg_attr(
-        not(feature = "gil-refs"),
-        deprecated(
-            since = "0.21.0",
-            note = "use `&Bound<'_, T>` instead for this function argument"
-        )
-    )]
-    pub fn extract_gil_ref(&self) {}
-}
-
-impl<T> NotAGilRef<T> {
-    pub fn extract_gil_ref(&self) {}
-    pub fn is_python(&self) {}
-}
-
-impl<T> std::ops::Deref for Extractor<T> {
-    type Target = NotAGilRef<T>;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
